Rename tables.py → model_data.py
